Moved Command History code into its own static class
