+++ VCF.Core/Registry/CommandRegistry.cs
@@ -31,8 +31,7 @@ public static class CommandRegistry
 		AssemblyCommandMap.Clear();
 		_converters.Clear();
 		_cache = new();
-		_commandHistory.Clear();
-		_loadedHistories.Clear();
+		CommandHistory.Reset();
 		_pendingCommands.Clear();
 	}
 
@@ -43,14 +42,14 @@ public static class CommandRegistry
 	// Store pending commands for selection
 	private static Dictionary<string, (string input, List<(CommandMetadata Command, object[] Args, string Error)> commands)> _pendingCommands = new();
 
-	private static Dictionary<ulong, List<(string input, CommandMetadata Command, object[] Args)>> _commandHistory = new();
-	private const int MAX_COMMAND_HISTORY = 10; // Store up to 10 past commands
-	
-	// Track which users have had their history loaded this session
-	private static HashSet<ulong> _loadedHistories = new();
-	
-	// Command history directory path
-	private static string HistoryDirectory => Path.Combine(Path.Combine(Paths.ConfigPath, PluginInfo.PLUGIN_NAME), "CommandHistory");
+	internal static CacheResult GetCommandFromCache(string input, string assemblyName = null)
+	{
+		if (assemblyName != null)
+		{
+			return _cache.GetCommandFromAssembly(input, assemblyName);
+		}
+		return _cache.GetCommand(input);
+	}
 
 	internal static bool CanCommandExecute(ICommandContext ctx, CommandMetadata command)
 	{
@@ -205,11 +204,7 @@ public static class CommandRegistry
 	public static CommandResult Handle(ICommandContext ctx, string input)
 	{
 		// Load command history for this user if it's their first command this session
-		var platformId = GetPlatformId(ctx);
-		if (platformId.HasValue && !_loadedHistories.Contains(platformId.Value))
-		{
-			LoadHistoryFromFile(ctx, platformId.Value);
-		}
+		CommandHistory.EnsureHistoryLoaded(ctx);
 
 		// Check if this is a command selection (e.g., .1, .2, etc.)
 		if (input.StartsWith(DEFAULT_PREFIX) && input.Length > 1)
@@ -229,8 +224,7 @@ public static class CommandRegistry
 
 		if (input.Trim().StartsWith(".!"))
 		{
-			HandleCommandHistory(ctx, input.Trim());
-			return CommandResult.Success;
+			return CommandHistory.HandleHistoryCommand(ctx, input.Trim(), Handle, ExecuteCommandWithArgs);
 		}
 
 		// Remove the prefix for processing
@@ -325,11 +319,12 @@ public static class CommandRegistry
 		if (successfulCommands.Count == 1)
 		{
 			var (command, commandArgs, _) = successfulCommands[0];
-			AddToCommandHistory(ctx, input, command, commandArgs);
+			CommandHistory.AddToHistory(ctx, input, command, commandArgs);
 			return ExecuteCommandWithArgs(ctx, command, commandArgs);
 		}
 
 		// Case 3: Multiple commands succeeded - store and ask user to select
+		var platformId = GetPlatformId(ctx);
 		var pendingKey = platformId.HasValue ? platformId.Value.ToString() : ctx.Name;
 		_pendingCommands[pendingKey] = (input, successfulCommands);
 
@@ -371,13 +366,13 @@ public static class CommandRegistry
 
 		var (command, args, _) = pendingCommands.commands[selectedIndex - 1];
 
-		AddToCommandHistory(ctx, pendingCommands.input, command, args);
+		CommandHistory.AddToHistory(ctx, pendingCommands.input, command, args);
 		var result = ExecuteCommandWithArgs(ctx, command, args);
 		_pendingCommands.Remove(pendingKey);
 		return result;
 	}
 
-	private static (bool Success, object[] Args, string Error) TryConvertParameters(ICommandContext ctx, CommandMetadata command, string[] args, string originalInput = null)
+	internal static (bool Success, object[] Args, string Error) TryConvertParameters(ICommandContext ctx, CommandMetadata command, string[] args, string originalInput = null)
 	{
 		var argCount = args?.Length ?? 0;
 		var paramsCount = command.Parameters.Length;
@@ -703,7 +698,7 @@ public static class CommandRegistry
 			return CommandResult.UsageError;
 		}
 
-		AddToCommandHistory(ctx, input, command, commandArgs);
+		CommandHistory.AddToHistory(ctx, input, command, commandArgs);
 		return ExecuteCommandWithArgs(ctx, command, commandArgs);
 	}
 
@@ -778,220 +773,6 @@ public static class CommandRegistry
 		return CommandResult.Success;
 	}
 
-	private static void AddToCommandHistory(ICommandContext ctx, string input, CommandMetadata command, object[] args)
-	{
-		var platformId = GetPlatformId(ctx);
-		if (!platformId.HasValue)
-		{
-			return; // Cannot save history for non-chat contexts
-		}
-
-		// Create the history list for this platform ID if it doesn't exist yet
-		if (!_commandHistory.TryGetValue(platformId.Value, out var history))
-		{
-			history = new List<(string input, CommandMetadata Command, object[] Args)>();
-			_commandHistory[platformId.Value] = history;
-		}
-
-		// Check if this exact command with same arguments already exists in history
-		for (int i = 0; i < history.Count; i++)
-		{
-			var historyEntry = history[i];
-
-			// Skip entries that couldn't be parsed at load
-			if (historyEntry.Command == null) continue;
-			
-			// Compare command metadata (same command method) and arguments
-			if (historyEntry.Command.Method == command.Method && 
-				historyEntry.Command.Attribute.Name == command.Attribute.Name &&
-				ArgsEqual(historyEntry.Args, args))
-			{
-				// Remove the existing duplicate
-				history.RemoveAt(i);
-				break; // Only remove the first match found as it should be the only one
-			}
-		}
-
-		// Add the new command to the beginning of the list
-		history.Insert(0, (input, command, args));
-
-		// Keep only the most recent MAX_COMMAND_HISTORY commands
-		if (history.Count > MAX_COMMAND_HISTORY)
-		{
-			history.RemoveAt(history.Count - 1);
-		}
-
-		// Save the updated history to file
-		SaveHistoryToFile(platformId.Value, history);
-	}
-
-	private static bool ArgsEqual(object[] args1, object[] args2)
-	{
-		if (args1 == null && args2 == null) return true;
-		if (args1 == null || args2 == null) return false;
-		if (args1.Length != args2.Length) return false;
-
-		// Skipping the first index as that is always the context
-		for (int i = 1; i < args1.Length; i++)
-		{
-			if (!Equals(args1[i], args2[i]))
-			{
-				return false;
-			}
-		}
-
-		return true;
-	}
-
-	private static void HandleCommandHistory(ICommandContext ctx, string input)
-	{
-		var platformId = GetPlatformId(ctx);
-		if (!platformId.HasValue)
-		{
-			ctx.SysReply($"{"[error]".Color(Color.Red)} Command history is not available for this context type.");
-			return;
-		}
-
-		// Remove the ".!" prefix
-		string command = input.Substring(2).Trim();
-
-		// Check if the command history exists for this platform ID
-		if (!_commandHistory.TryGetValue(platformId.Value, out var history) || history.Count == 0)
-		{
-			ctx.SysReply($"{"[error]".Color(Color.Red)} No command history available.");
-			return;
-		}
-
-		// Handle .! list or .! l commands
-		if (command == "list" || command == "l")
-		{
-			var sb = new StringBuilder();
-			sb.AppendLine("Command history:");
-
-			for (int i = 0; i < history.Count; i++)
-			{
-				sb.AppendLine($"{(i + 1).ToString().Color(Color.Gold)}. {history[i].input.Color(Color.Command)}");
-			}
-
-			ctx.SysPaginatedReply(sb);
-			return;
-		}
-
-		// Handle .! # to execute a specific command by number
-		if (int.TryParse(command, out int index) && index > 0 && index <= history.Count)
-		{
-			var selectedCommand = history[index - 1];
-			ctx.SysReply($"Executing command {index.ToString().Color(Color.Gold)}: {selectedCommand.input.Color(Color.Command)}");
-			
-			// If Command and Args are available (successfully parsed), use them directly
-			if (selectedCommand.Command != null && selectedCommand.Args != null)
-			{
-				ExecuteCommandWithArgs(ctx, selectedCommand.Command, selectedCommand.Args);
-			}
-			else
-			{
-				// Fall back to re-parsing if command wasn't successfully parsed during load
-				Handle(ctx, selectedCommand.input);
-			}
-			return;
-		}
-
-		// If just .! is provided, execute the most recent command
-		if (string.IsNullOrWhiteSpace(command))
-		{
-			var mostRecent = history[0];
-			ctx.SysReply($"Repeating most recent command: {mostRecent.input.Color(Color.Command)}");
-			
-			// If Command and Args are available (successfully parsed), use them directly
-			if (mostRecent.Command != null && mostRecent.Args != null)
-			{
-				ExecuteCommandWithArgs(ctx, mostRecent.Command, mostRecent.Args);
-			}
-			else
-			{
-				// Fall back to re-parsing if command wasn't successfully parsed during load
-				Handle(ctx, mostRecent.input);
-			}
-			return;
-		}
-
-		// Invalid command
-		ctx.SysReply($"{"[error]".Color(Color.Red)} Invalid command history selection. Use {".! list".Color(Color.Command)} to see available commands or {".! #".Color(Color.Command)} to execute a specific command.");
-	}
-
-	private static void SaveHistoryToFile(ulong platformId, List<(string input, CommandMetadata Command, object[] Args)> history)
-	{
-		try
-		{
-			if (!Directory.Exists(HistoryDirectory))
-			{
-				Directory.CreateDirectory(HistoryDirectory);
-			}
-
-			string filePath = Path.Combine(HistoryDirectory, $"{platformId}.txt");
-			var inputsOnly = history.Select(h => h.input).ToArray();
-			File.WriteAllLines(filePath, inputsOnly);
-		}
-		catch (Exception ex)
-		{
-			Log.Error($"Failed to save command history for platform ID {platformId}: {ex.Message}");
-		}
-	}
-
-	private static void LoadHistoryFromFile(ICommandContext ctx, ulong platformId)
-	{
-		try
-		{
-			string filePath = Path.Combine(HistoryDirectory, $"{platformId}.txt");
-
-			if (!File.Exists(filePath))
-			{
-				return; // No history file exists
-			}
-
-			var lines = File.ReadAllLines(filePath);
-			if (lines.Length == 0)
-			{
-				return; // Empty file
-			}
-
-			var reconstructedHistory = new List<(string input, CommandMetadata Command, object[] Args)>();
-
-			// Process lines in reverse to maintain chronological order (most recent first)
-			lines.Reverse();
-			foreach (var input in lines)
-			{
-				try
-				{
-					// Parse the command to get CommandMetadata and Args
-					var (command, args) = ParseCommandForHistory(ctx, input);
-					if (command != null && args != null)
-					{
-						reconstructedHistory.Add((input, command, args));
-					}
-					else
-					{
-						// If parsing fails, still add the input for display purposes
-						reconstructedHistory.Add((input, null, null));
-					}
-				}
-				catch (Exception)
-				{
-					// If individual command parsing fails, add with null values
-					reconstructedHistory.Add((input, null, null));
-				}
-			}
-
-			_commandHistory[platformId] = reconstructedHistory;
-			_loadedHistories.Add(platformId);
-		}
-		catch (Exception ex)
-		{
-			Log.Error($"Failed to load command history for platform ID {platformId}: {ex.Message}");
-			_loadedHistories.Add(platformId); // Mark as loaded to avoid repeated attempts
-		}
-	}
-
 	private static ulong? GetPlatformId(ICommandContext ctx)
 	{
 		if (ctx is ChatCommandContext chatCtx)
@@ -999,77 +780,6 @@ public static class CommandRegistry
 			return chatCtx.User.PlatformId;
 		}
 		return null;
-	}
-
-	private static (CommandMetadata command, object[] args) ParseCommandForHistory(ICommandContext ctx, string input)
-	{
-		try
-		{
-			// Ensure the command starts with the prefix
-			if (!input.StartsWith(DEFAULT_PREFIX))
-			{
-				return (null, null);
-			}
-
-			// Remove the prefix for processing
-			string afterPrefix = input.Substring(DEFAULT_PREFIX.Length);
-
-			// Check if this could be an assembly-specific command
-			string assemblyName = null;
-			string commandInput = input;
-
-			int spaceIndex = afterPrefix.IndexOf(' ');
-			if (spaceIndex > 0)
-			{
-				string potentialAssemblyName = afterPrefix.Substring(0, spaceIndex);
-
-				// Check if this could be a valid assembly name
-				bool isValidAssembly = AssemblyCommandMap.Keys.Any(assemblyName =>
-					assemblyName.Equals(potentialAssemblyName, StringComparison.OrdinalIgnoreCase));
-
-				if (isValidAssembly)
-				{
-					assemblyName = potentialAssemblyName;
-					commandInput = "." + afterPrefix.Substring(spaceIndex + 1);
-				}
-			}
-
-			// Get command(s) based on input
-			CacheResult matchedCommand = null;
-			if (assemblyName != null)
-			{
-				matchedCommand = _cache.GetCommandFromAssembly(commandInput, assemblyName);
-			}
-			if (matchedCommand == null || !matchedCommand.IsMatched)
-			{
-				matchedCommand = _cache.GetCommand(input);
-			}
-
-			var commands = matchedCommand.Commands;
-
-			if (!matchedCommand.IsMatched || !commands.Any())
-			{
-				return (null, null);
-			}
-
-			// Try to find the first command that can be parsed successfully
-			foreach (var (command, cmdArgs) in commands)
-			{
-				if (!CanCommandExecute(ctx, command)) continue;
-
-				var (success, commandArgs, error) = TryConvertParameters(ctx, command, cmdArgs, input);
-				if (success)
-				{
-					return (command, commandArgs);
-				}
-			}
-
-			return (null, null);
-		}
-		catch (Exception)
-		{
-			return (null, null);
-		}
 	}
 
 	public static void UnregisterConverter(Type converter)
